Use `u16`
Upper bits are reserved.

+++ kernel/src/device/pci/xhci/register/runtime_base_registers.rs
@@ -29,8 +29,8 @@ impl<'a> RuntimeBaseRegisters<'a> {
 #[derive(Debug)]
 pub struct EventRingSegmentTableSizeRegister(u32);
 impl EventRingSegmentTableSizeRegister {
-    pub fn set(&mut self, val: u32) {
-        self.0 = val
+    pub fn set(&mut self, val: u16) {
+        self.0 = u32::from(val)
     }
 }
